# packages/contracts/openapi/agent-runtime.openapi.yaml
openapi: 3.1.0
info:
  title: Insight Agent Runtime API
  version: 0.1.0
  description: >
    Agent Runtime 服务的最小 OpenAPI 契约。该文档只定义 runtime contract surface，
    不代表后端 handler、worker、SSE 或真实模型/工具执行已经实现。
paths:
...
  /analysis-runs:
    post:
      summary: 创建分析运行
      description: >
        创建一次新的 AnalysisRun 契约对象，并把 run 绑定到既有 conversationId。
        该接口只创建 `created / intake` foundation run，不写 Message、RunEvent 或报告对象。
      operationId: createAnalysisRun
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAnalysisRunRequest"
            example:
              workspaceId: workspace-northstar-retail-china
              userId: user-zoe
              analysisTaskId: analysis-task-6d7a1a7097b34752ae5d5cb9b390bc0e
              conversationId: conversation-0c948fb61b8b4dfc95d81568c41c5a3f
      responses:
        "201":
          description: AnalysisRun 创建成功后的正式契约响应。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisRun"
              example:
                runId: analysis-run-2c1ee63bc1a549e68c4780fe7a5bfb9a
                workspaceId: workspace-northstar-retail-china
                userId: user-zoe
                analysisTaskId: analysis-task-6d7a1a7097b34752ae5d5cb9b390bc0e
                status: created
                phase: intake
                outcome: null
                waitingFor: null
                createdAt: "2026-06-05T03:08:30Z"
                validatingAt: null
                queuedAt: null
                startedAt: null
                waitingSince: null
                timeoutAt: null
                cancelRequestedAt: null
                cancellingAt: null
                completedAt: null
                failedAt: null
                cancelledAt: null
                expiredAt: null
                rejectedAt: null
                terminalReason: null
                failureCode: null
                retryable: true
                retryOfRunId: null
                originalRunId: null
        "404":
          $ref: "#/components/responses/RuntimeNotFound"
        "409":
          $ref: "#/components/responses/RuntimeMismatch"
  /analysis-runs/{runId}:
    get:
      summary: 获取分析运行详情
      operationId: getAnalysisRun
      parameters:
        - $ref: "#/components/parameters/RunId"
      responses:
        "200":
          description: AnalysisRun 详情。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisRun"
              example:
                runId: analysis-run-2c1ee63bc1a549e68c4780fe7a5bfb9a
                workspaceId: workspace-northstar-retail-china
                userId: user-zoe
                analysisTaskId: analysis-task-6d7a1a7097b34752ae5d5cb9b390bc0e
                status: created
                phase: intake
                outcome: null
                waitingFor: null
                createdAt: "2026-06-05T03:08:30Z"
                validatingAt: null
                queuedAt: null
                startedAt: null
                waitingSince: null
                timeoutAt: null
                cancelRequestedAt: null
                cancellingAt: null
                completedAt: null
                failedAt: null
                cancelledAt: null
                expiredAt: null
                rejectedAt: null
                terminalReason: null
                failureCode: null
                retryable: true
                retryOfRunId: null
                originalRunId: null
        "404":
          $ref: "#/components/responses/RuntimeNotFound"
  /analysis-runs/{runId}/dispatch:
    post:
      summary: 推进分析运行进入最小 dispatch foundation
      description: >
        把 `created/intake` 的 AnalysisRun 推进到 `queued/queueing`，
        并创建一条真实 ExecutionAttempt，为后续 #159 Worker / Queue / Dispatch 承接。
      operationId: dispatchAnalysisRun
      parameters:
        - $ref: "#/components/parameters/RunId"
      responses:
        "202":
          description: 已完成最小 lifecycle 推进和 ExecutionAttempt 创建后的 AnalysisRun。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisRun"
              example:
                runId: analysis-run-2c1ee63bc1a549e68c4780fe7a5bfb9a
                workspaceId: workspace-northstar-retail-china
                userId: user-zoe
                analysisTaskId: analysis-task-6d7a1a7097b34752ae5d5cb9b390bc0e
                status: queued
                phase: queueing
                outcome: null
                waitingFor: null
                createdAt: "2026-06-05T03:08:30Z"
                validatingAt: "2026-06-05T03:08:35Z"
                queuedAt: "2026-06-05T03:08:35Z"
                startedAt: null
                waitingSince: null
                timeoutAt: null
                cancelRequestedAt: null
                cancellingAt: null
                completedAt: null
                failedAt: null
                cancelledAt: null
                expiredAt: null
                rejectedAt: null
                terminalReason: null
                failureCode: null
                retryable: true
                retryOfRunId: null
                originalRunId: null
        "404":
          $ref: "#/components/responses/RuntimeNotFound"
        "409":
          $ref: "#/components/responses/RuntimeInvalidState"
  /analysis-runs/{runId}/conversation:
    get:
      summary: 获取运行关联会话
      description: 返回当前 AnalysisRun 关联的 Conversation 契约对象。
      operationId: getAnalysisRunConversation
      parameters:
        - $ref: "#/components/parameters/RunId"
      responses:
        "200":
          description: AnalysisRun 当前绑定的 Conversation。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
              example:
                conversationId: conversation-0c948fb61b8b4dfc95d81568c41c5a3f
                workspaceId: workspace-northstar-retail-china
                userId: user-zoe
                analysisTaskId: analysis-task-6d7a1a7097b34752ae5d5cb9b390bc0e
                currentRunId: analysis-run-2c1ee63bc1a549e68c4780fe7a5bfb9a
                title: 收入增速异常
                status: active
                createdAt: "2026-06-05T03:08:20Z"
                updatedAt: "2026-06-05T03:08:30Z"
        "404":
          $ref: "#/components/responses/RuntimeNotFound"
  /analysis-runs/{runId}/events:
    get:
      summary: 获取运行事件列表
      operationId: listRunEvents
      parameters:
        - $ref: "#/components/parameters/RunId"
      responses:
        "200":
          description: RunEvent 列表。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunEventList"
              example:
                items:
                  - eventId: event-analysis-q2-revenue-gap-user-input
                    runId: analysis-q2-revenue-gap
                    eventType: run.created
                    status: succeeded
                    phase: intake
                    sequence: 0
                    actor: analysis_runtime_stub
                    occurredAt: "2026-06-05T11:08:00+08:00"
                    summary: 记录当前用户问题，并为后续运行绑定上下文。
                    parentEventId: null
                    refType: null
                    refId: null
                    errorCode: null
                    errorMessage: null
                    nodeName: run.created
                    agentName: analysis-agent
                    toolName: null
                    startedAt: "2026-06-05T11:08:00+08:00"
                    completedAt: "2026-06-05T11:08:00+08:00"
                  - eventId: event-analysis-q2-revenue-gap-feedback-waiting
                    runId: analysis-q2-revenue-gap
                    eventType: run.completed
                    status: succeeded
                    phase: delivery
                    sequence: 7
                    actor: analysis_runtime_stub
                    occurredAt: "2026-06-05T11:15:00+08:00"
                    summary: 当前运行完成，等待用户继续追问或进入报告入口。
                    parentEventId: null
                    refType: null
                    refId: null
                    errorCode: null
                    errorMessage: null
                    nodeName: run.completed
                    agentName: analysis-agent
                    toolName: null
                    startedAt: "2026-06-05T11:15:00+08:00"
                    completedAt: "2026-06-05T11:15:00+08:00"
        "501":
          $ref: "#/components/responses/NotImplementedRuntimeRoute"
  /analysis-runs/{runId}/tool-calls:
    get:
      summary: 获取工具调用列表
      operationId: listToolCalls
      parameters:
        - $ref: "#/components/parameters/RunId"
      responses:
        "200":
          description: ToolCall 列表。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToolCallList"
        "501":
          $ref: "#/components/responses/NotImplementedRuntimeRoute"
  /analysis-runs/{runId}/model-calls:
    get:
      summary: 获取模型调用列表
      operationId: listModelCalls
      parameters:
        - $ref: "#/components/parameters/RunId"
      responses:
        "200":
          description: ModelCall 列表。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelCallList"
        "501":
          $ref: "#/components/responses/NotImplementedRuntimeRoute"
  /analysis-runs/{runId}/source-evidence:
    get:
      summary: 获取来源证据列表
      operationId: listSourceEvidence
      parameters:
        - $ref: "#/components/parameters/RunId"
      responses:
        "200":
          description: SourceEvidence 列表。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourceEvidenceList"
        "501":
          $ref: "#/components/responses/NotImplementedRuntimeRoute"
  /analysis-runs/{runId}/reports:
    get:
      summary: 获取报告列表
      operationId: listReports
      parameters:
        - $ref: "#/components/parameters/RunId"
      responses:
        "200":
          description: Report 列表。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportList"
        "501":
          $ref: "#/components/responses/NotImplementedRuntimeRoute"
  /analysis-runs/{runId}/execution-attempts:
    get:
      summary: 获取执行尝试列表
      operationId: listExecutionAttempts
      parameters:
        - $ref: "#/components/parameters/RunId"
      responses:
        "200":
          description: ExecutionAttempt 列表。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecutionAttemptList"
              example:
                items:
                  - attemptId: attempt-7d60f811c53d447bb1c7e2c1d6445dd5
                    runId: analysis-run-2c1ee63bc1a549e68c4780fe7a5bfb9a
                    attemptNumber: 1
                    workerId: worker-runtime-dispatch-foundation
                    leaseId: lease-f307d2263a264e3593f394c8b9bb7887
                    status: leased
                    leaseAcquiredAt: "2026-06-05T03:08:35Z"
                    leaseExpiresAt: "2026-06-05T03:28:35Z"
                    heartbeatAt: null
                    releasedAt: null
                    failureCode: null
                    failureMessage: null
        "404":
          $ref: "#/components/responses/RuntimeNotFound"
  /analysis-runs/{runId}/approvals:
    get:
      summary: 获取审批请求列表
      operationId: listApprovalRequests
      parameters:
        - $ref: "#/components/parameters/RunId"
      responses:
        "200":
          description: ApprovalRequest 列表。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovalRequestList"
        "501":
          $ref: "#/components/responses/NotImplementedRuntimeRoute"
  /analysis-runs/{runId}/cancel:
    post:
      summary: 取消分析运行
      operationId: cancelAnalysisRun
      parameters:
        - $ref: "#/components/parameters/RunId"
      responses:
        "202":
          description: 已记录取消请求后的正式契约响应示例。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisRun"
        "501":
          $ref: "#/components/responses/NotImplementedRuntimeRoute"
  /analysis-runs/{runId}/retry:
    post:
      summary: 基于旧运行创建 retry
      operationId: retryAnalysisRun
      parameters:
        - $ref: "#/components/parameters/RunId"
      responses:
        "202":
          description: 已创建新的 retry AnalysisRun 的正式契约响应示例。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisRun"
        "501":
          $ref: "#/components/responses/NotImplementedRuntimeRoute"
  /analysis-runs/{runId}/approvals/{approvalId}/decision:
    post:
      summary: 提交审批决定
      operationId: decideApprovalRequest
      parameters:
        - $ref: "#/components/parameters/RunId"
        - $ref: "#/components/parameters/ApprovalId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalDecisionRequest"
            example:
              status: granted
              decisionReason: Finance lead approved the retry.
      responses:
        "200":
          description: 更新后的 ApprovalRequest。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovalRequest"
        "501":
          $ref: "#/components/responses/NotImplementedRuntimeRoute"
components:
  responses:
    NotImplementedRuntimeRoute:
      description: Runtime route stub is registered; real implementation is pending.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/NotImplementedErrorResponse"
          example:
            errorCode: NOT_IMPLEMENTED
            message: Runtime route stub is registered; real implementation is pending.
    RuntimeNotFound:
      description: Requested runtime foundation object was not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/RuntimeErrorResponse"
          example:
            errorCode: NOT_FOUND
            message: "AnalysisTask not found: analysis-task-missing"
    RuntimeMismatch:
      description: Request chain mismatched the persisted runtime foundation objects.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/RuntimeErrorResponse"
          example:
            errorCode: MISMATCH
            message: Conversation.analysisTaskId does not match request.analysisTaskId
    RuntimeInvalidState:
      description: Requested lifecycle transition is not allowed for the persisted AnalysisRun state.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/RuntimeErrorResponse"
          example:
            errorCode: INVALID_STATE
            message: AnalysisRun must be created/intake before dispatch.
  parameters:
    RunId:
      name: runId
      in: path
      required: true
      schema:
        type: string
      description: AnalysisRun 的 canonical runId。
    ApprovalId:
      name: approvalId
      in: path
      required: true
      schema:
        type: string
      description: ApprovalRequest 的 canonical approvalId。
...
  schemas:
...
    NotImplementedErrorResponse:
      type: object
      additionalProperties: false
      required:
        - errorCode
        - message
      properties:
        errorCode:
          type: string
          enum:
            - NOT_IMPLEMENTED
        message:
          type: string
    RuntimeErrorResponse:
      type: object
      additionalProperties: false
      required:
        - errorCode
        - message
      properties:
        errorCode:
          type: string
          enum:
            - NOT_FOUND
            - MISMATCH
            - INVALID_STATE
        message:
          type: string
...
    CreateAnalysisRunRequest:
      type: object
      additionalProperties: false
      required:
        - workspaceId
        - userId
        - analysisTaskId
        - conversationId
      properties:
        workspaceId:
          type: string
        userId:
          type: string
        analysisTaskId:
          type: string
        conversationId:
          type: string
    ApprovalDecisionRequest:
      type: object
      additionalProperties: false
      required:
        - status
      properties:
        status:
          type: string
          enum:
            - granted
            - denied
            - cancelled
        decisionReason:
          type:
            - string
            - "null"
...
    AnalysisRun:
      $ref: "../schemas/analysis/analysis-run.schema.json"
    Conversation:
      $ref: "../schemas/analysis/conversation.schema.json"
...
    RunEvent:
      $ref: "../schemas/analysis/run-event.schema.json"
    ToolCall:
      $ref: "../schemas/analysis/tool-call.schema.json"
    ModelCall:
      $ref: "../schemas/analysis/model-call.schema.json"
    SourceEvidence:
      $ref: "../schemas/analysis/source-evidence.schema.json"
    Report:
      $ref: "../schemas/reports/report.schema.json"
    ExecutionAttempt:
      $ref: "../schemas/analysis/execution-attempt.schema.json"
    ApprovalRequest:
      $ref: "../schemas/analysis/approval-request.schema.json"
    RunEventList:
      type: object
      additionalProperties: false
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/RunEvent"
...
    ToolCallList:
      type: object
      additionalProperties: false
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ToolCall"
    ModelCallList:
      type: object
      additionalProperties: false
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ModelCall"
    SourceEvidenceList:
      type: object
      additionalProperties: false
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/SourceEvidence"
    ReportList:
      type: object
      additionalProperties: false
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Report"
    ExecutionAttemptList:
      type: object
      additionalProperties: false
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ExecutionAttempt"
    ApprovalRequestList:
      type: object
      additionalProperties: false
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ApprovalRequest"